+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Table of Contents
 
 ### Bug Fixes
 
+* Fix a rare `java.nio.file.FileSystemAlreadyExistsException` when accessing filters concurrently.
+
 ### New Features
 
 * Metadata differences are now filterable via the recheck.ignore.