+++ CHANGELOG.md
@@ -27,14 +27,20 @@ Table of Contents
 
 ### Bug Fixes
 
-* Fix 1-click-maintenance not respecting the key of the `AttributeDifference`. 
-
 ### New Features
 
 ### Improvements
 
 
 --------------------------------------------------------------------------------
+
+
+[1.11.1] (2020-06-10)
+---------------------
+
+### Bug Fixes
+
+* Fix 1-click-maintenance not respecting the key of the `AttributeDifference`. 
 
 
 [1.11.0] (2020-05-15)